extract PaginatedConversationIDList
